requirements.txt
# Core
torch>=2.1.0
transformers>=4.40.0
datasets>=2.18.0
tokenizers>=0.15.0
safetensors>=0.4.0
accelerate>=0.27.0

# Training
trl>=0.8.0
peft>=0.10.0

# Hessian analysis
pyhessian>=0.1

# RepE / analysis
scikit-learn>=1.4.0
numpy>=1.26.0
scipy>=1.12.0

# Logging and tracking
wandb>=0.16.0

# Config
pyyaml>=6.0

# Progress bars
tqdm>=4.66.0
